fix: handle empty success parameter in GetCronHistory
This commit fixes an issue in the `GetCronHistory` method where an empty `success` parameter would cause an error.

The change modifies the `success` parameter to be a pointer to a boolean. This allows the code to differentiate between a `success` parameter that is explicitly set to `false` and a `success` parameter that is not provided at all. If the `success` parameter is not provided, the code will retrieve all cron history entries, regardless of their success status. If the `success` parameter is provided but is invalid, the endpoint returns a 400 error.